chore(developer): rename build/ to buildClasses/
Our dev env mostly assumes that build/ is for build artifacts, so this
just sidesteps issues with code search, etc.

The other glaring example of this is resources/build/ which is a _much_
more painful change, given how many scripts reference it. So not
changing that for now.
