+++ developer/src/kmc/src/commands/buildClasses/BuildActivity.ts
