+++ developer/src/kmc/src/commands/buildClasses/BuildKmnKeyboard.ts
