+++ developer/src/kmc/src/commands/buildClasses/BuildModel.ts
