+++ developer/src/kmc/src/commands/buildClasses/BuildPackage.ts
