+++ developer/src/kmc/src/commands/buildClasses/BuildProject.ts
