+++ developer/src/kmc/src/commands/buildClasses/buildActivities.ts
